requirements.txt
qrcode
pyzbar
tqdm
Pillow
mss